moved MultiLogMessageWaitStrategy to core
